OIL-98: restructure test utils
