+++ test/specs/userview/userview_handle-optin.spec.js
@@ -1,5 +1,5 @@
-import {handleOptIn,} from '../../../src/scripts/userview/userview_modal.js';
-import {deleteAllCookies, loadFixture, removeOilLayerAndConfig} from '../../utils.js';
+import { handleOptIn, } from '../../../src/scripts/userview/userview_modal.js';
+import { loadFixture } from '../../test-utils/utils_fixtures.js';
 import * as CoreConfig from '../../../src/scripts/core/core_config.js';
 import * as UserviewConfig from '../../../src/scripts/userview/userview_config.js';
 import * as UserViewPrivacy from '../../../src/scripts/userview/userview_privacy.js';
@@ -7,7 +7,8 @@ import * as CoreUtils from '../../../src/scripts/core/core_utils.js';
 import * as UserviewOptIn from '../../../src/scripts/userview/userview_optin.js'
 import * as CorePoi from '../../../src/scripts/core/core_poi.js';
 import * as CoreCookies from '../../../src/scripts/core/core_cookies.js';
-import {EVENT_NAME_POI_OPT_IN, EVENT_NAME_SOI_OPT_IN, PRIVACY_FULL_TRACKING, PRIVACY_MINIMUM_TRACKING} from '../../../src/scripts/core/core_constants';
+import { EVENT_NAME_POI_OPT_IN, EVENT_NAME_SOI_OPT_IN, PRIVACY_FULL_TRACKING, PRIVACY_MINIMUM_TRACKING } from '../../../src/scripts/core/core_constants';
+import { deleteAllCookies, removeOilLayerAndConfig } from '../../test-utils/utils_reset';
 
 describe('the userview modal handles optin clicks on', () => {
